+++ CHANGELOG.rst
@@ -5,6 +5,10 @@ Changelog
 latest - will be *v0.9.0*
 -------------------------
 
+- New routine:
+  - ``emg3d.utils.Fourier``, a class do handle Fourier-transform related stuff
+    for time-domain modelling. See the example notebooks for its usage.
+
 - Utilities:
 
   - ``Fields`` and returned receiver-arrays (``EMArray``) both have amplitude